+++ miataru/miataru/views/Mac/Mac_OnboardingContainerView.swift
@@ -34,6 +34,9 @@ struct Mac_OnboardingContainerView: View {
                 pages.append(AnyView(iPhone_4_OnboardingLocationHistoryView()))
             }
             if settings.trackAndReportLocation {
+                pages.append(AnyView(iPhone_9_OnboardingUnknownVisitorAlertsView()))
+            }
+            if settings.trackAndReportLocation {
                 pages.append(AnyView(iPhone_6_OnboardingDeviceKeyView()))
             }
             if settings.trackAndReportLocation {